Move location slice out to its new temporary home
